more udp support

+++ src/P2p/LevinProtocol.h
@@ -97,7 +97,7 @@ private:
   bool readStrict(uint8_t* ptr, size_t size);
   void writeStrict(const uint8_t* ptr, size_t size);
   System::TcpConnection& m_conn;
-  int udpPort;
+  uint16_t udpPort;
 };
 
 }
